requirements.txt
asgiref==3.7.2
certifi==2025.4.26
charset-normalizer==3.4.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-environ==0.12.0
django-import-export==3.3.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==23.0.0
idna==3.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.2
pillow==10.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.4.4
sslcommerz_python_api==1.0.1
tablib==3.5.0
tzdata==2024.1
urllib3==2.4.0
whitenoise==6.9.0
xlrd==2.0.1
xlwt==1.3.0
